Added check so as to not overwrite existing files

+++ ninite-scrape.sh
@@ -1,3 +1,5 @@
 while read i; do
-	wget --no-check-certificate -O "wget/ninite $i installer.exe" "https://ninite.com/$i/ninite.exe"
+	if [ ! -f "wget/ninite $i installer.exe" ]; then
+		wget --no-check-certificate -O "wget/ninite $i installer.exe" "https://ninite.com/$i/ninite.exe"
+	fi
 done <ninite-all.txt
